iot-basics/requirements.txt
boto3
paho-mqtt
